+++ RELEASE_NOTES.md
@@ -3,7 +3,7 @@
 * Added parameter 'reference_product' to findActivity
 * Improved cache of impacts : trakc per-database modificiation time. Put bio proxies in a separate database.
 * Added possibility to use target activity as key in updateExchange()
-
+* add 'strict_mode' in Settings to forbid any creation or update of activity in background database
 
 # **